requirements.txt
hydra-core==1.1.1
transformers==4.15.0
datasets==1.17.0
overrides==6.1.0
sacrebleu==2.0.0
